pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.11.2",
    "balldontlie>=0.1.6",
    "beautifulsoup4>=4.14.3",
    "email-validator>=2.3.0",
    "flask>=3.1.2",
    "flask-bcrypt>=1.0.1",
    "flask-cors>=6.0.2",
    "flask-limiter>=4.1.1",
    "flask-login>=0.6.3",
    "flask-sqlalchemy>=3.1.1",
    "google-auth>=2.48.0",
    "gunicorn>=25.0.3",
    "itsdangerous>=2.2.0",
    "kagglehub>=0.4.2",
    "lightgbm>=4.6.0",
    "numpy>=2.4.2",
    "oauthlib>=3.3.1",
    "pandas>=3.0.0",
    "psycopg2-binary>=2.9.11",
    "pyjwt>=2.11.0",
    "requests>=2.32.5",
    "resend>=2.21.0",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.0",
    "sqlalchemy>=2.0.46",
    "stripe>=14.3.0",
    "urllib3>=2.6.3",
    "werkzeug>=3.1.5",
    "xgboost>=3.1.3",
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
